# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/repos": {
            "get": {
                "description": "分页查询仓库列表，支持按状态筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仓库管理"
                ],
                "summary": "获取仓库列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "状态筛选(pending/cloning/ready/failed)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/repos/batch": {
            "post": {
                "description": "批量添加多个Git仓库，异步克隆到本地",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仓库管理"
                ],
                "summary": "批量添加仓库",
                "parameters": [
                    {
                        "description": "仓库URL列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.AddReposRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.AddReposResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/repos/{id}": {
            "get": {
                "description": "根据ID获取仓库详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仓库管理"
                ],
                "summary": "获取仓库详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Repository"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定仓库",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仓库管理"
                ],
                "summary": "删除仓库",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/repos/{id}/branches": {
            "get": {
                "description": "获取指定仓库的所有分支",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仓库管理"
                ],
                "summary": "获取仓库分支列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/repos/{id}/reset": {
            "post": {
                "description": "异步重置仓库到最新状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仓库管理"
                ],
                "summary": "重置仓库",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Task"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/repos/{id}/switch-branch": {
            "post": {
                "description": "异步切换仓库到指定分支",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仓库管理"
                ],
                "summary": "切换仓库分支",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "分支名称",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "branch": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Task"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/repos/{id}/update": {
            "post": {
                "description": "异步拉取仓库最新代码(git pull)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仓库管理"
                ],
                "summary": "更新仓库",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "仓库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Task"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Repository": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_branch": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "has_credentials": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "last_commit_hash": {
                    "type": "string"
                },
                "last_pull_at": {
                    "type": "string"
                },
                "local_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "pending/cloning/ready/failed",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.Task": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_ms": {
                    "description": "计算字段",
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "parameters": {
                    "description": "JSON string",
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "repo_id": {
                    "type": "integer"
                },
                "result": {
                    "description": "JSON string",
                    "type": "string"
                },
                "retry_count": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "task_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "service.AddRepoFailure": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "service.AddRepoResult": {
            "type": "object",
            "properties": {
                "repo_id": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "service.AddReposRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "可选的认证信息",
                    "type": "string"
                },
                "repos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.RepoInput"
                    }
                },
                "username": {
                    "description": "可选的认证信息",
                    "type": "string"
                }
            }
        },
        "service.AddReposResponse": {
            "type": "object",
            "properties": {
                "failed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.AddRepoFailure"
                    }
                },
                "succeeded": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.AddRepoResult"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "service.RepoInput": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}